Reorganize user routes
