+++ mon-pix/mirage/routes/users/index.js
@@ -4,6 +4,7 @@ import getPixScore from './get-pix-score';
 import getScorecards from './get-scorecards';
 import getUserCampaignParticipations from './get-user-campaign-participations';
 import resetScorecard from './reset-scorecard';
+import { Response } from 'ember-cli-mirage';
 
 export default function index(config) {
   config.post('/users');
@@ -13,4 +14,22 @@ export default function index(config) {
   config.get('/users/:id/scorecards', getScorecards);
   config.get('/users/:id/campaign-participations', getUserCampaignParticipations);
   config.post('/users/:userId/competences/:competenceId/reset', resetScorecard);
+  config.patch('/users/:id/password-update', (schema, request) => {
+    const body = JSON.parse(request.requestBody);
+    const user =  schema.users.find(request.params.id);
+
+    const demand = schema.passwordResetDemands.findBy({ temporaryKey: request.queryParams['temporary-key'] });
+    if (user.email !== demand.email) {
+      return new Response(401);
+    } else {
+      user.update({ password: body.data.attributes.password });
+      return new Response(204);
+    }
+  });
+
+  config.patch('/users/:id/remember-user-has-seen-assessment-instructions', (schema, request) => {
+    const user =  schema.users.find(request.params.id);
+    user.hasSeenAssessmentInstructions = true;
+    return user;
+  });
 }
